skel/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Bcrypt==0.7.1
flask-restful-swagger==0.20.1
